+++ packages/bff/src/index.ts
@@ -1,3 +1,2 @@
 export { AppModule } from "./bootstrap/app.module";
-export { BootstrapService } from "./bootstrap/bootstrap.service";
 export { startBffServer } from "./bootstrap/main";