+++ packages/datasource/src/infra/index.ts
@@ -1,2 +1,4 @@
 export * from "./config";
 export * from "./postgres/postgres.adapter";
+export * from "./postgres/orders-mutation.adapter";
+export * from "./postgres/org-scope.adapter";